refactor(abi): split contract! codegen into per-concern submodules
The single 1.3k-line codegen.rs became `codegen/mod.rs` (the
`generate_contract` orchestrator) plus one submodule per concern:

- naming  - casing / identifier helpers (the leaf dependency)
- structs - struct support resolution, struct codegen, ABI tuple types
- methods - arg specs, method fns, builders, lifecycle setters, the
            unsupported-method scan
- client  - the client struct, its impl, network constructors
- events  - ARC-28 event enum + decoder
- state   - global-state read accessors
- deploy  - the deploy constructor

Cross-module helpers are `pub(super)`; everything else stays private.
Tests move next to the code they cover. Pure code movement -- no
behaviour change; the macro unit tests and the integration suite are
unchanged and green.
